+++ src/Application/Command/Meeting/EvaluationTimeoutMessage.php
@@ -2,7 +2,6 @@
 
 namespace Proximum\Vimeet\Application\Command\Meeting;
 
-use Proximum\Vimeet\Domain\Model\Contact;
 use Proximum\Vimeet\Domain\Model\Meeting;
 use Proximum\Vimeet\Domain\Model\User;
 
@@ -13,15 +12,11 @@ class EvaluationTimeoutMessage
 
     private int $meetingId;
     private int $fromUserId;
-    /** @var int[] $contactIds */
-    private array $contactIds;
 
-    /** @param Contact[] $contacts */
-    public function __construct(Meeting $meeting, User $fromUser, array $contacts)
+    public function __construct(Meeting $meeting, User $fromUser)
     {
         $this->meetingId = $meeting->getId();
         $this->fromUserId = $fromUser->getId();
-        $this->contactIds = array_map(fn (Contact $c) => $c->getContact()->getId(), $contacts);
     }
 
     public function getMeetingId(): int
@@ -32,11 +27,5 @@ class EvaluationTimeoutMessage
     public function getFromUserId(): int
     {
         return $this->fromUserId;
-    }
-
-    /** @return int[] */
-    public function getContactIds(): array
-    {
-        return $this->contactIds;
     }
 }