+++ src/Application/Query/Networking/NetworkingNotAccessibleException.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Proximum\Vimeet\Application\Query\Networking;
+
+
+class NetworkingNotAccessibleException extends \RuntimeException
+{
+    public $message = 'The networking is close or sheet has no access to networking';
+}