+++ src/Application/Query/Participant/CardListViewQueryHandler.php
@@ -3,32 +3,28 @@
 namespace Proximum\Vimeet\Application\Query\Participant;
 
 use Proximum\Vimeet\Application\View\Participant\CardListView;
+use Proximum\Vimeet\Domain\KeyDates\Checker\NetworkingAccessChecker;
 
 class CardListViewQueryHandler
 {
-    /**
-     * @var CardViewQueryHandler
-     */
-    private $cardViewQueryHandler;
+    private CardViewQueryHandler $cardViewQueryHandler;
+    private NetworkingAccessChecker $networkingAccessChecker;
 
-    /**
-     * @param CardViewQueryHandler $cardViewQueryHandler
-     */
-    public function __construct(CardViewQueryHandler $cardViewQueryHandler)
-    {
+    public function __construct(
+        CardViewQueryHandler $cardViewQueryHandler,
+        NetworkingAccessChecker $networkingAccessChecker
+    ) {
         $this->cardViewQueryHandler = $cardViewQueryHandler;
+        $this->networkingAccessChecker = $networkingAccessChecker;
     }
 
-    /**
-     * @param CardListViewQuery $cardListViewQuery
-     *
-     * @return CardListView
-     */
-    public function handle(CardListViewQuery $cardListViewQuery)
+    public function handle(CardListViewQuery $cardListViewQuery): CardListView
     {
         $participants = $cardListViewQuery->sheet->getParticipants();
-        $user         = $cardListViewQuery->user;
+        $user = $cardListViewQuery->user;
         $cardListView = new CardListView();
+
+        $showMeetOnline = $this->networkingAccessChecker->isSheetAllowedToAccess($cardListViewQuery->sheet);
 
         foreach ($participants as $participant) {
             if ($cardListViewQuery->editable) {
@@ -37,7 +33,7 @@ class CardListViewQueryHandler
                 $editable = false;
             }
 
-            $cardViewQuery = new CardViewQuery($participant, $cardListViewQuery->locale, $editable);
+            $cardViewQuery = new CardViewQuery($participant, $cardListViewQuery->locale, $editable, $showMeetOnline);
 
             $cardListView->cardViews[$participant->getId()] = $this->cardViewQueryHandler->handle($cardViewQuery);
         }