+++ src/Ui/Bundle/EventBundle/TemplateSnippet/NetworkingTemplateSnippet.php
@@ -5,7 +5,7 @@ namespace Proximum\Vimeet\Ui\Bundle\EventBundle\TemplateSnippet;
 
 
 use Proximum\Vimeet\Application\Adapter\QueryBusInterface;
-use Proximum\Vimeet\Application\Query\Networking\ClosedNetworkingException;
+use Proximum\Vimeet\Application\Query\Networking\NetworkingNotAccessibleException;
 use Proximum\Vimeet\Application\Query\Networking\GetSnippetQuery;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
@@ -40,7 +40,7 @@ class NetworkingTemplateSnippet
 
         try {
             $getSnippetView = $this->queryBus->handle(new GetSnippetQuery($sheet, $user));
-        } catch (ClosedNetworkingException $e) {
+        } catch (NetworkingNotAccessibleException $e) {
             return ' ';
         }